Main Functionality
This code allows for a majority of the specifications for database management. For demonstration purposes, the submission of all scores and all run times have been collapsed into two methods and are accessed in tandem. This functionality will be altered for use with the GUI. **IMPORTANT** Despite all attempts, the MySQL connector module would not connect to any IDE or other python editor I could find so this code may still contain logic errors